Make first set of suggested changes (file structure & goofs)
