+++ src/plugins/drawing/components/vector-type-button.tsx
@@ -1,11 +1,8 @@
 import classNames from "classnames";
 import React from "react";
 import { observer } from "mobx-react";
-import { VectorType } from "./vector-palette";
-import LineToolIcon from "../assets/line-icon.svg";
-import SingleArrowIcon from "../assets/line-single-arrow-icon.svg";
-import DoubleArrowIcon from "../assets/line-double-arrow-icon.svg";
-import { ToolbarSettings } from "../model/drawing-basic-types";
+import { ToolbarSettings, VectorType } from "../model/drawing-basic-types";
+import { VectorTypeIcon } from "../objects/vector";
 
 interface IProps {
   vectorType: VectorType;
@@ -28,26 +25,4 @@ export const VectorTypeButton = observer(
   );
 });
 
-interface IVectorTypeIconProps {
-  vectorType: VectorType;
-  settings: ToolbarSettings;
-}
 
-export function VectorTypeIcon ({ vectorType, settings }: IVectorTypeIconProps) {
-  // SVG attributes to use when drawing the icon.
-  // Note that the arrowheads are filled with the stroke color, we don't use settings.fill for this
-  const attributes = {
-    stroke: settings.stroke, 
-    fill: settings.stroke, // uses stroke for fill
-    strokeWidth: settings.strokeWidth,
-    strokeDasharray: settings.strokeDashArray
-  };
-  switch(vectorType) {
-    case VectorType.line:
-      return <LineToolIcon {...attributes} />;
-    case VectorType.singleArrow:
-      return <SingleArrowIcon {...attributes} />;
-    case VectorType.doubleArrow:
-      return <DoubleArrowIcon {...attributes} />;
-  }
-}
