+++ src/plugins/drawing/model/drawing-basic-types.ts
@@ -1,4 +1,3 @@
-import { VectorType } from "../components/vector-palette";
 
 export interface Point { x: number; y: number; }
 
@@ -7,6 +6,32 @@ export interface BoundingBox {
   se: Point;
   start?: Point;
   end?: Point;
+}
+
+export enum VectorType {
+  line = "line",
+  singleArrow = "arrow",
+  doubleArrow = "doublearrow"
+}
+
+// Possible decorations for the start and end of the vector.  Default is no decoration.
+export enum VectorEndShape {
+  triangle = "triangle"
+}
+
+// Return a two-element list of [head shape, tail shape] for the given VectorType constant.
+export function endShapesForVectorType(vectorType?: VectorType) {
+  if (!vectorType) {
+    return [undefined, undefined];
+  }
+  switch (vectorType) {
+    case VectorType.line:
+      return [undefined, undefined];
+    case VectorType.singleArrow:
+      return [VectorEndShape.triangle, undefined];
+    case VectorType.doubleArrow:
+      return [VectorEndShape.triangle, VectorEndShape.triangle];
+  }
 }
 
 export interface ToolbarSettings {
@@ -24,8 +49,3 @@ export const DefaultToolbarSettings: ToolbarSettings = {
   strokeWidth: 2,
   vectorType: VectorType.line
 };
-
-// Possible decorations for the start and end of the vector.  Default is no decoration.
-export enum VectorEndShape {
-  triangle = "triangle"
-}
